Fix: creating empty temp file when withoutTempFiles() mode is enabled
Fix for issue when TesseractOCR is executed with option withoutTempFiles()
but temporary output file is created while getOuptutFile() is called
inside FirendlyErrors::checkCommandExecution() during checking a source
(file or stdout) for valid result output.

This issue can cause an occurance of many undeleted empty ocr* files in
temp directory and finally can lead to a lack of inodes.

Old behaviour: empty ocr* file is created on each launch of
TesseractOCR::withoutTempFiles() and TesseractOcr::run().

New behaviour: empty ocr* file is not created in this case.